Adding verifications and changing features

+++ bibliotecaPDF/DTOs/SelfUserDTO.cs
@@ -5,7 +5,7 @@ namespace bibliotecaPDF.DTOs;
 public record SelfUserDTO(string Name, string Email, string Telefone, double CapacityUsed)
 {
     
-    public SelfUserDTO(User user) : this(user.Name, user.Email, user.PhoneNumber, (user.ByteAmounts / user.CapacityPlan.BytesCapacity) * 100)
+    public SelfUserDTO(User user) : this(user.Name, user.Email, user.PhoneNumber, ((double)user.ByteAmounts / (double)user.CapacityPlan.BytesCapacity) * 100)
     {
     }
 };
